refactor(cfd-math): Decompose dg/operators/mod.rs — extract DGOperator leaf module
Part of ATLAS-ARCH-007.
